+++ arches/app/src/components/ControlledListManager/ControlledListTable.vue
