+++ arches/app/src/components/ControlledListManager/ControlledListsAll.vue
